scripts/requirements.txt
openai
qdrant-client
beautifulsoup4
markdown-it-py